Fetch intrinsic gas and update CeloBlockEnv (#23)
* core_contracts: Add get_intrinsic_gas

* core_contracts: Add update_block_env

* core_contracts: Avoid fetching currencies twice

* core_contracts: use EvmTr to avoid .0.0 access

* core_contracts: Import revm::Database

to make the trait constraints a bit shorter.

* core_contracts: Move update_fee_currencies to CeloBlockEnv

This is more similar to the related L1BlockInfo::try_fetch.

The `clone` in the function is not really necessary since we replace the
original anyway, but otherwise it doesn't fit this API. We could also
return the exchange_rates and intrinsic_gas instead, but that would make
the usage a bit less nice.