Call update_fee_currencies, more specific traits
To call update_fee_currencies, we must use a CeloEvm. In addition to
making the trait constraints more specific, I also make the CeloEvm more
specific by only taking two type constraints. This is not necessary, but
it makes the code more readable and we don't need the additional
abstraction.

I should have separated the CeloEvm simplification from the other
change, but that was not clear to me while making everything compile
again.

It is not clear if set_block is the best place to call this and what the
best error handling in this case would be. But I'm optimistic that this
is a step in the right direction.