core_contracts: Improve calling infrastructure (#38)
* Use transact_system_call for core contracts

This functions does less work and is not tied to our transaction types
wrapper. It also seems to be made for calls like this.

* Restore tx set in EVM after system call

We don't want to modify the EVM object when doing system calls.

* core_contracts: more detailed error output

* Revert even on failure

Clean up the order of operations, so that we always revert state changes
and restore the old tx before checking for errors. This avoids
duplication and makes it harder to forget cleaning up the state.