Fix block executor example
We were missing the holocene and isthmus hardfork times.

This meant that the executor was not correctly constructing the base fee
params from the parent header and also not setting the base fee params
in the extra data of the constructed header.

Also updated the executor assertion message to make it clear which value
is the expected and which is the actual.