More specific constraints + FeeCurrencyContext (#37)
* Fee currency context

* Apply formatter

* Apply clippy --fix

* More specific traits

To call update_fee_currencies, we must use a CeloEvm. In addition to
making the trait constraints more specific, I also make the CeloEvm more
specific by only taking two type constraints. This is not necessary, but
it makes the code more readable and we don't need the additional
abstraction.

I should have separated the CeloEvm simplification from the other
change, but that was not clear to me while making everything compile
again.

* Constrain CeloHandler to CeloEVM

We don't need to be more generic and this makes it easier to interact
with Celo specific constructs like the FeeCurrencyContext.

* fixup! Constrain CeloHandler to CeloEVM

---------

Co-authored-by: Gastón Ponti <gaston.ponti@clabs.co>