Fix block executor example for isthmus (#40)
We were missing the holocene and isthmus hardfork times in our baklava config.

This meant that the executor was not correctly constructing the base fee
params from the parent header and also not setting the base fee params
in the extra data of the constructed header.

Also updated the executor assertion message to make it clear which value
is the expected and which is the actual.

Additionally renamed the saved execution fixtures which lets us easily
understand the significance of the the specific blocks.