Wip mostly compiling TriedDBProvider
Except for the header fetching, realy header fetching shouldn't be part
of the TriedDB interface because it's not to do with the trie.

Unfortunately I've realised that the header fetching requires a provider
since headers are stored in the ancient db in geth after some time.

So for now I'm thinking use a provider and a leveldb db here.

Aslo I'm starting to question the current approach for the
ExecutorTestFixtureCreator it's wrapping the OP one but really only
using the fields of it. Also the ExecutorTestFixtureCreator implements
TriedDB on itslef and then passes itself to its dependents which is just
ugly.