fix: track unsuccessfully verified blocks as well
Before we did not insert the block into the highest-verified block
tracker if the block header had a mismatch.

However we DO want to track this block, because it had been verified,
just that the outcome was unsuccessful. Otherwise the persistence
feature would not work as expected.

If the block should be re-verified, e.g. after a fix has been conducted,
then the verifier should be re-run with a fresh state-file.