feat: add OpenTelemetry log-tracing and metrics (#56)
* feat: add OpenTelemetry log-tracing and metrics

* chore: cleanup unused dependencies

* chore: fmt and lint fixes

* fix: add missing 'derive' feature in clap

* chore: omit OTLP exporter instead of using Option type

* chore: log OTel debug tracing

* chore: put OTLP log message after initialization

* fix: use the rt-tokio compatible OTLP batch_exporter

Before, the simple exporter seems to have been blocking the main thread
and thus hindering the rest of the program to execute.

* chore: add log-target to fmt-layer

* feat: refactor metrics and expose highest-verified-block metric

* chore: add recommended OTel resource attributes

* fix: track unsuccessfully verified blocks as well

Before we did not insert the block into the highest-verified block
tracker if the block header had a mismatch.

However we DO want to track this block, because it had been verified,
just that the outcome was unsuccessful. Otherwise the persistence
feature would not work as expected.

If the block should be re-verified, e.g. after a fix has been conducted,
then the verifier should be re-run with a fresh state-file.

* chore: move and rename `celo-kona-utils` to `celo-otel`

* fix(celo-otel): return error when verbosity-level is 0