Set the l1 block base fee scalars to zero
On Alfajores we initially didn't have these set to zero but our cost
function always returned zero. Then we decided to set these to zero and
not change the cost function. We eventually set the fee scalars to zero
in this block 32786341.

This is a hacky fix but should suffice to test alfjores excution before
we shut it down in 2 weeks.