Remove custom CeloSystemCallTx trait (#93)
* Remove custom CeloSystemCallTx trait

Fewer traits, more readability!

* Use upstream SystemCallTx trait from revm instead of custom trait.
* Move CELO_SYSTEM_ADDRESS constant to constants.rs module.
* Delete celo_system_tx.rs file as it's no longer needed.
* Add concrete new_system_tx_with_gas_limit method to CeloTransaction<TxEnv>.

* Use fixed gas limit for new_system_tx_with_caller

I'm not sure how relevant the 30_000_000 actually is. I took it as "an arbitrary value that is high enough to do all core contract calls", but it might matter in some theoretical cases. I'll put the 30_000_000 back to not break something during the refactor, but it would be nice to clarify this at some point.

See https://github.com/celo-org/celo-kona/pull/93/files#r2473044393