Patch revm with our forked version of v97
This ensures that we are using the same revm version as in our
op-succinct repo. The v97 version fixed an execution bug that would have
led to an execution mismatch between the op-stack and our
succinct-stack.

Also updates dependencies to point at the versions exported by the base
of our forked revm.