CIP-64: actual GASPRICE opcode fix (#107)
The previous fix was not correct and only accidentally made the tests
pass. The GASPRICE opcodes fetches the effective gas price and passes in
the base fee in native tokens. We work around this by calculating the
correct effective gas price in advance and ignoring the passed in base
fee.