Validate base_fee when effective_gas_price is called
In 76bad58b4b153c2dc898362651fdb63cd2f5f156, we introduced a workaround
for GASPRICE calculation in CIP-64 txs. This assume that the
precalculated effective gas price is still valid. To notice problem
cases, we also store the base_fee at the same time and compare with he
base fee passed to effective_gas_price, which should be used in the
calculation, but can't due to the lack of exchange rates.