Better CIP-64 GASPRICE opcode test
The previous test case often passed even with wrong implementations
because it only noticed if a single `if` condition on the gas price
triggers. Instead, the actual gas price return value should be part of
the tested result to notice any deviations from the expected value.

This is achieved by emitting the result in an event/log so that it
becomes part of the receipt. Full bytecode explanation:

| Offset | Bytecode | Opcode       | Stack (after)   | Description                              |
|--------|----------|--------------|-----------------|------------------------------------------|
| 0x00   | `3a`     | `GASPRICE`   | `[gasPrice]`    | Pushes the tx's gas price onto the stack |
| 0x01   | `60 00`  | `PUSH1 0x00` | `[0, gasPrice]` | Pushes memory offset 0                   |
| 0x03   | `52`     | `MSTORE`     | `[]`            | Stores gasPrice at memory[0:32]          |
| 0x04   | `60 20`  | `PUSH1 0x20` | `[32]`          | Pushes 32 (0x20) - the size in bytes     |
| 0x06   | `60 00`  | `PUSH1 0x00` | `[0, 32]`       | Pushes memory offset 0                   |
| 0x08   | `a0`     | `LOG0`       | `[]`            | Emits anonymous log from memory[0:32]    |

As intended, the test behaves like this:
```
fail
cd86616 Fix CIP-64 gas price opcode (#106)
fail
9d02406 CIP-64: actual GASPRICE opcode fix (#107)
pass (the missing revert has no effect on the returned gas price)
7be19b3 Revert bad fix for GASPRICE opcode (#109)
pass
```
So if we had the test before doing the changes, we would have correctly
rejected the wrong fix.