FeeCurrencyContext error handling consistency with op-geth (#104)
* Consistency with op-geth

* Creating the fee currency context can't fail, only discard currencies.
* Exchange rate can't be zero (neither numerator nor denominator)
* Skip currencies on any call error

Closes https://github.com/celo-org/celo-blockchain-planning/issues/1280

* Ensure that currencies are completely loaded or not at all

If only one of exchange rate or intrinsic gas is available, the fee
currency must be discarded.

* Combine exchange rate and intrinsic gas into FeeCurrencyInfo

This way, states where only exchange rate of intrinsic gas have been
read can't be represented. This avoids the consistency check and makes
it obvious that both must always be present for each fee currency.

* refactor: keep get_currencies internal to core_contracts.rs